Rename queryserver-client to query-sever-client (kebab-case)
