+++ extensions/ql-vscode/src/query-server/query-server-client.ts
