+++ extensions/ql-vscode/src/query-server/run-queries.ts
@@ -2,7 +2,7 @@ import { CancellationToken } from "vscode";
 import { ProgressCallback } from "../progress";
 import * as messages from "../pure/new-messages";
 import { QueryOutputDir } from "../run-queries-shared";
-import * as qsClient from "./queryserver-client";
+import * as qsClient from "./query-server-client";
 import { CoreQueryResults, CoreQueryTarget } from "../queryRunner";
 import { Logger } from "../common";
