+++ extensions/ql-vscode/test/vscode-tests/cli-integration/new-query.test.ts
@@ -2,11 +2,11 @@ import { join, basename } from "path";
 import { dirSync } from "tmp";
 import { CancellationTokenSource } from "vscode-jsonrpc";
 import * as messages from "../../../src/pure/new-messages";
-import * as qsClient from "../../../src/query-server/queryserver-client";
+import * as qsClient from "../../../src/query-server/query-server-client";
 import * as cli from "../../../src/cli";
 import { CellValue } from "../../../src/pure/bqrs-cli-types";
 import { describeWithCodeQL } from "../cli";
-import { QueryServerClient } from "../../../src/query-server/queryserver-client";
+import { QueryServerClient } from "../../../src/query-server/query-server-client";
 import { extLogger, ProgressReporter } from "../../../src/common";
 import { QueryResultType } from "../../../src/pure/new-messages";
 import { ensureTestDatabase, getActivatedExtension } from "../global.helper";
